Primary Objective 2
Display Zika Reports on Map. This objective covers a ton of ground. Lots of it is stuff student's probably haven't been exposed to, like Feature, FeatureCollection, GeoJSONSerializer. Luckily all of this stuff was provided for them in the airwaze project. The writeup should help them figure some of it out, but the students' objective should be to have a strong grasp of how everything works together for the Layer to be displayed on the map correctly. A very benefical exercise to do as a group is to draw the flow of data. CSV -> report table -> Report object -> Feature -> FeatureCollection. One of the gotchas in this that students' haven't seen before is how to add the feature if, and only if the location geometry is present. I imagine lots of students' will be tripped up here, because nothing in the writeup explicitly mentions it.